requirements.txt
streamlit>=1.30.0
huggingface_hub>=0.23.0
python-dotenv>=1.0.0
